# perf: Speed up `getTargets`