<!--
Before making a PR, please read our contributing guidelines
https://github.com/babel/babel/blob/main/CONTRIBUTING.md

Please note that the Babel Team requires two approvals before merging most PRs.

For issue references: Add a comma-separated list of a [closing word](https://help.github.com/articles/closing-issues-via-commit-messages/) followed by the ticket number fixed by the PR. (it should be underlined in the preview if done correctly)

If you are making a change that should have a docs update: submit another PR to https://github.com/babel/website
-->

| Q                        | A <!--(Can use an emoji 👍) -->
| ------------------------ | ---
| Fixed Issues?            | `Fixes #1, Fixes #2` <!-- remove the (`) quotes and write "Fixes" before the number to link the issues -->
| Patch: Bug Fix?          |
| Major: Breaking Change?  |
| Minor: New Feature?      |
| Tests Added + Pass?      | Yes
| Documentation PR Link    | <!-- If only readme change, add `[skip ci]` to your commits -->
| Any Dependency Changes?  |
| License                  | MIT

<!-- Describe your changes below in as much detail as possible -->

ref: #15214 

I chose to cache `resolveTargets` instead of `getTargets`, because the parameters of `getTargets` contain dynamic paths, which will bring additional burden to maintenance and usage.

`resolveTargets`
```
PS F:\babel\benchmark\babel-helper-compilation-targets> node .\getTargets.mjs
baseline#getTargets last1: 73_804 ops/sec ±0.88% (0.014ms)
baseline#getTargets last100: 2_141 ops/sec ±0.81% (0.467ms)
baseline#getTargets chrome 49 ie 11: 111_644 ops/sec ±0.81% (0.009ms)
current#getTargets last1: 291_856 ops/sec ±0.55% (0.003ms)
current#getTargets last100: 301_740 ops/sec ±1.47% (0.003ms)
current#getTargets chrome 49 ie 11: 544_397 ops/sec ±0.87% (0.002ms)
```

`getTargets`
```
PS F:\babel\benchmark\babel-helper-compilation-targets> node .\getTargets.mjs
baseline#getTargets last1: 72_310 ops/sec ±2.6% (0.014ms)
baseline#getTargets last100: 2_137 ops/sec ±0.5% (0.468ms)
baseline#getTargets chrome 49 ie 11: 115_129 ops/sec ±0.4% (0.009ms)
current#getTargets last1: 877_577 ops/sec ±0.53% (0.001ms)
current#getTargets last100: 856_543 ops/sec ±0.56% (0.001ms)
current#getTargets chrome 49 ie 11: 1_294_732 ops/sec ±0.61% (0.001ms)
```